+++ CHANGES.md
@@ -9,7 +9,8 @@
 * Tweak CWL format (#24)
 * Use micromamba entry point in Docker image (#26)
 * Allow setting of CWL workflow ID (#29)
-* Add in-notebook configuration interface (#30)
+* Support in-notebook configuration of workflow ID, environment file,
+  and container image tag (#30, #33)
 * Support writing of stage-out STAC by notebook (#32)
 
 ## Changes in 0.1.0